Refactor more stuff.

+++ Engine/src/mini/material/Technique.java
@@ -150,8 +150,8 @@ public class Technique {
      *
      * @return the sort ID for this technique instance.
      */
-    public int getSortId() {
-        int hash = 17;
+    public long getSortId() {
+        long hash = 17;
         hash = hash * 23 + def.getSortId();
         hash = hash * 23 + paramDefines.hashCode();
         return hash;
